Fix Visual Studio build.
Tested: Builds on Visual Studio and Mac now.

+++ src/components/render_3d_text.cpp
@@ -66,7 +66,8 @@ const mat4 Render3dTextComponent::CalculateAnimationTransform(
   }
 
   const RenderMeshData* rendermesh_data = Data<RenderMeshData>(entity);
-  const int num_mesh_bones = rendermesh_data->mesh->num_bones();
+  const int num_mesh_bones =
+      static_cast<int>(rendermesh_data->mesh->num_bones());
 
   if (num_mesh_bones > 1 && num_mesh_bones > animation_bone) {
     return mat4::FromAffineTransform(
